# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RC Track Café API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type BookingStatus = typeof BookingStatus[keyof typeof BookingStatus];


export const BookingStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  cancelled: 'cancelled',
} as const;

export interface Booking {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  date: string;
  time: string;
  experienceType: string;
  /** @nullable */
  specialRequests?: string | null;
  status: BookingStatus;
  createdAt: string;
}

export interface BookingInput {
  /** @minLength 1 */
  firstName: string;
  /** @minLength 1 */
  lastName: string;
  /** @minLength 1 */
  email: string;
  phone?: string;
  date: string;
  time: string;
  /** @minLength 1 */
  experienceType: string;
  specialRequests?: string;
}

export type BookingUpdateStatus = typeof BookingUpdateStatus[keyof typeof BookingUpdateStatus];


export const BookingUpdateStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  cancelled: 'cancelled',
} as const;

export interface BookingUpdate {
  status?: BookingUpdateStatus;
  firstName?: string;
  lastName?: string;
  email?: string;
  phone?: string;
  date?: string;
  time?: string;
  experienceType?: string;
  specialRequests?: string;
}

export interface Service {
  id: number;
  name: string;
  description: string;
  priceFrom: number;
  priceUnit: string;
  category: string;
  featured?: boolean;
}

export type MenuItemCategory = typeof MenuItemCategory[keyof typeof MenuItemCategory];


export const MenuItemCategory = {
  espresso: 'espresso',
  filter: 'filter',
  cold: 'cold',
  food: 'food',
  specialty: 'specialty',
} as const;

export interface MenuItem {
  id: number;
  name: string;
  description: string;
  price: number;
  category: MenuItemCategory;
  featured?: boolean;
  sortOrder?: number;
}

export type MenuItemInputCategory = typeof MenuItemInputCategory[keyof typeof MenuItemInputCategory];


export const MenuItemInputCategory = {
  espresso: 'espresso',
  filter: 'filter',
  cold: 'cold',
  food: 'food',
  specialty: 'specialty',
} as const;

export interface MenuItemInput {
  /** @minLength 1 */
  name: string;
  description: string;
  price: number;
  category: MenuItemInputCategory;
  featured?: boolean;
  sortOrder?: number;
}

export type MenuItemUpdateCategory = typeof MenuItemUpdateCategory[keyof typeof MenuItemUpdateCategory];


export const MenuItemUpdateCategory = {
  espresso: 'espresso',
  filter: 'filter',
  cold: 'cold',
  food: 'food',
  specialty: 'specialty',
} as const;

export interface MenuItemUpdate {
  name?: string;
  description?: string;
  price?: number;
  category?: MenuItemUpdateCategory;
  featured?: boolean;
  sortOrder?: number;
}

export interface GalleryItem {
  id: number;
  title: string;
  category: string;
  /** @nullable */
  imageUrl?: string | null;
}

export interface ContactInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  email: string;
  subject?: string;
  /** @minLength 1 */
  message: string;
}

export interface ContactMessage {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  subject?: string | null;
  message: string;
  createdAt: string;
}

export interface VenueStats {
  totalBookings: number;
  confirmedBookings: number;
  pendingBookings: number;
  totalTracks: number;
  memberCount: number;
  coffeeVarieties: number;
  yearsOpen: number;
}