# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: RC Track Café API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /menu:
    get:
      operationId: listMenuItems
      tags: [menu]
      summary: List all menu items
      responses:
        "200":
          description: Menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"
    post:
      operationId: createMenuItem
      tags: [menu]
      summary: Create a menu item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
        "400":
          description: Bad request
  /menu/{id}:
    patch:
      operationId: updateMenuItem
      tags: [menu]
      summary: Update a menu item
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuItemUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
        "404":
          description: Not found
    delete:
      operationId: deleteMenuItem
      tags: [menu]
      summary: Delete a menu item
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
components:
  schemas:
...
    MenuItem:
      type: object
      required: [id, name, description, price, category]
      properties:
        id:          { type: integer }
        name:        { type: string }
        description: { type: string }
        price:       { type: number }
        category:    { type: string, enum: [espresso, filter, cold, food, specialty] }
        featured:    { type: boolean }
        sortOrder:   { type: integer }
    MenuItemInput:
      type: object
      required: [name, description, price, category]
      properties:
        name:        { type: string, minLength: 1 }
        description: { type: string }
        price:       { type: number }
        category:    { type: string, enum: [espresso, filter, cold, food, specialty] }
        featured:    { type: boolean }
        sortOrder:   { type: integer }
    MenuItemUpdate:
      type: object
      properties:
        name:        { type: string }
        description: { type: string }
        price:       { type: number }
        category:    { type: string, enum: [espresso, filter, cold, food, specialty] }
        featured:    { type: boolean }
        sortOrder:   { type: integer }
...