# lib/api-zod/src/generated/types/menuItemInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RC Track Café API
 * OpenAPI spec version: 0.1.0
 */
import type { MenuItemInputCategory } from './menuItemInputCategory';

export interface MenuItemInput {
  /** @minLength 1 */
  name: string;
  description: string;
  price: number;
  category: MenuItemInputCategory;
  featured?: boolean;
  sortOrder?: number;
}