# lib/api-zod/src/generated/types/menuItemUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RC Track Café API
 * OpenAPI spec version: 0.1.0
 */
import type { MenuItemUpdateCategory } from './menuItemUpdateCategory';

export interface MenuItemUpdate {
  name?: string;
  description?: string;
  price?: number;
  category?: MenuItemUpdateCategory;
  featured?: boolean;
  sortOrder?: number;
}